requirements.txt
pandas
numpy
matplotlib
scikit-learn
python-dotenv
kagglehub[pandas-datasets]
google-genai
groq